migrating to Swift 3

+++ OpenStack Summit/CoreSummit/Link.swift
@@ -18,7 +18,7 @@ public struct Link: PresentationMaterial {
     
     public var featured: Bool
     
-    public var order: Int
+    public var order: Int64
     
     public var link: String
